+++ src/db/transactions.rs
@@ -47,8 +47,20 @@ where
         Ok(())
     }
 
+    /// Set a `key` to a value in the database if the key does not exist.
+    pub fn set_nx_raw<'k, 'v>(&'v self, key: impl AsRef<[u8]>, val: &'v [u8]) -> Result<bool> {
+        let res = self.column.set_nx(key, val)?;
+        Ok(res)
+    }
+
+    /// Set a `key` to a serialized value in the database if the key does not exist.
+    pub fn set_nx<'k, 'v>(&'v self, key: &'k <Key>::EItem, val: &'v <Val>::EItem) -> Result<bool>
+    where
+        Key: BytesEncode<'k>,
+        Val: BytesEncode<'v>;
+
     /// Get the serialized `val` from the database by `key`.
-    pub fn get<'k, 'v>(&self, key: &'k <Key>::EItem) -> Result<<Val>::DItem>
+    pub fn get<'k, 'v>(&self, key: &'k <Key>::EItem) -> Result<Option<<Val>::DItem>>
     where
         Key: BytesEncode<'k>,
         Val: BytesDecodeOwned;